The user group of drawn places can now be deleted too. A custom name of that group is saved.
