+++ src/model/place.ts
@@ -24,25 +24,16 @@
 
 import * as  GeoJSON from 'geojson';
 import { defaultMemoize } from 'reselect';
-import { default as OlWktFormat } from "ol/format/WKT";
-import { default as OlGeomPoint } from "ol/geom/Point";
 import { default as OlGeoJSONFormat } from "ol/format/GeoJSON";
-import { default as OlFeature } from "ol/Feature";
 import { default as OlGeometry } from "ol/geom/Geometry";
-
-import { CsvOptions, defaultCsvOptions } from "./user-place/csv";
-import { defaultGeoJsonOptions, GeoJsonOptions } from "./user-place/geojson";
-import { defaultWktOptions, WktOptions } from "./user-place/wkt";
 import { newId } from "../util/id";
 import { getUserPlaceColorName } from '../config';
-import { parseCsv } from "../util/csv";
-import i18n from "../i18n";
 
 export const USER_ID_PREFIX = "user-";
 
 // The ID of the user place group that hold all drawn places.
 // This place group will always exist.
-export const USER_DRAWING_PLACE_GROUP_ID = USER_ID_PREFIX + "drawing";
+export const USER_DRAWN_PLACE_GROUP_ID = USER_ID_PREFIX + "drawing";
 
 /**
  * A place is a GeoJSON feature with a mandatory string ID.
@@ -249,298 +240,3 @@ export function findPlaceInPlaceGroups(placeGroups: PlaceGroup[], placeId: strin
     }
     return null;
 }
-
-let LAST_PLACE_GROUP_ID_CSV = 0;
-let LAST_PLACE_GROUP_ID_GEO_JSON = 0;
-let LAST_PLACE_GROUP_ID_WKT = 0;
-
-let LAST_PLACE_LABEL_ID_CSV = 0;
-let LAST_PLACE_LABEL_ID_GEO_JSON = 0;
-let LAST_PLACE_LABEL_ID_WKT = 0;
-
-
-export function getUserPlacesFromCsv(text: string, options: CsvOptions): PlaceGroup[] {
-    let table = parseCsv(text, options);
-    if (table.length < 2) {
-        throw new Error(i18n.get('Missing header line in CSV'));
-    }
-
-    for (let v of table[0]) {
-        if (typeof v !== 'string' || v === '') {
-            throw new Error(i18n.get('Invalid header line in CSV'));
-        }
-    }
-
-    const headerRow: string[] = table[0].map(v => v as string);
-    const headerRowLC = headerRow.map(v => v.toLowerCase());
-    const numColumns = headerRow.length;
-
-    for (let row of table) {
-        if (row.length !== numColumns) {
-            throw new Error(i18n.get('All rows must have same length'));
-        }
-    }
-
-    const columnNameToIndex = getColumnNameToIndex(headerRowLC);
-    const groupNameIndex = findColumnIndex(columnNameToIndex, options.groupNames);
-    const labelNameIndex = findColumnIndex(columnNameToIndex, options.labelNames);
-    const xNameIndex = findColumnIndex(columnNameToIndex, options.xNames);
-    const yNameIndex = findColumnIndex(columnNameToIndex, options.yNames);
-    let geometryNameIndex = findColumnIndex(columnNameToIndex, options.geometryNames);
-    const forceGeometry = options.forceGeometry;
-
-    if (forceGeometry || xNameIndex < 0 || yNameIndex < 0 || xNameIndex === yNameIndex) {
-        if (geometryNameIndex < 0) {
-            throw new Error(i18n.get('No geometry column(s) found'));
-        }
-    } else {
-        geometryNameIndex = -1;
-    }
-
-    let groupPrefix = options.groupPrefix.trim();
-    if (groupPrefix === '') {
-        groupPrefix = defaultCsvOptions.groupPrefix;
-    }
-
-    let labelPrefix = options.labelPrefix.trim();
-    if (labelPrefix === '') {
-        labelPrefix = defaultCsvOptions.labelPrefix;
-    }
-
-    let group: string = '';
-    if (groupNameIndex === -1) {
-        const groupId = ++LAST_PLACE_GROUP_ID_CSV;
-        group = `${groupPrefix}${groupId}`;
-    }
-
-    const wktFormat = new OlWktFormat();
-
-    const placeGroups: { [group: string]: PlaceGroup } = {};
-    let rowIndex = 1;
-    let numPlaceGroups = 0;
-    let color = getUserPlaceColorName(0);
-    for (; rowIndex < table.length; rowIndex++) {
-        const dataRow = table[rowIndex];
-
-        if (groupNameIndex >= 0) {
-            group = `${dataRow[groupNameIndex]}`;
-        }
-
-        let placeGroup = placeGroups[group];
-        if (!placeGroup) {
-            placeGroup = newUserPlaceGroup(group, []);
-            placeGroups[group] = placeGroup;
-            color = getUserPlaceColorName(numPlaceGroups);
-            numPlaceGroups++;
-        }
-
-        let geometry = null;
-        if (geometryNameIndex >= 0) {
-            const wkt = dataRow[geometryNameIndex];
-            if (typeof wkt === 'string') {
-                try {
-                    geometry = wktFormat.readGeometry(text);
-                } catch (e) {
-                    // will handle error below
-                }
-            }
-        } else {
-            const x = dataRow[xNameIndex];
-            const y = dataRow[yNameIndex];
-            if (typeof x === 'number' && Number.isFinite(x) && typeof y === 'number' && Number.isFinite(y)) {
-                geometry = new OlGeomPoint([x, y]);
-            }
-        }
-        if (geometry === null) {
-            throw new Error(i18n.get(`Invalid geometry in data row ${rowIndex}`));
-        }
-
-        const geoJsonProps: { [k: string]: number | string | boolean | null } = {};
-        dataRow.forEach((propertyValue, index) => {
-            if (index !== xNameIndex && index !== yNameIndex && index !== geometryNameIndex) {
-                const propertyName = headerRow[index];
-                geoJsonProps[propertyName] = propertyValue;
-            }
-        });
-
-        let label: string;
-        if (labelNameIndex >= 0) {
-            label = `${dataRow[labelNameIndex]}`;
-        } else {
-            const labelId = ++LAST_PLACE_LABEL_ID_CSV;
-            label = `${labelPrefix}${labelId}`;
-        }
-
-        if (!geoJsonProps['color'])
-            geoJsonProps['color'] = color;
-        if (!geoJsonProps['label'])
-            geoJsonProps['label'] = label;
-        if (!geoJsonProps['source'])
-            geoJsonProps['source'] = 'CSV';
-
-        placeGroup.features.push(newUserPlace(geometry, geoJsonProps));
-    }
-
-    return Object.getOwnPropertyNames(placeGroups).map(group => placeGroups[group]);
-}
-
-export function getUserPlacesFromGeoJson(text: string, options: GeoJsonOptions): PlaceGroup[] {
-    const groupNames = options.groupNames;
-    const groupNamesSet = new Set(
-        groupNames.split(',')
-            .map(name => name.trim().toLowerCase())
-            .filter(name => name !== '')
-    )
-    let groupPrefix = options.groupPrefix.trim();
-    if (groupPrefix === '') {
-        groupPrefix = defaultGeoJsonOptions.groupPrefix;
-    }
-
-    const labelNames = options.labelNames;
-    const labelNamesSet = new Set(
-        labelNames.split(',')
-            .map(name => name.trim().toLowerCase())
-            .filter(name => name !== '')
-    )
-    let labelPrefix = options.labelPrefix.trim();
-    if (labelPrefix === '') {
-        labelPrefix = defaultGeoJsonOptions.labelPrefix;
-    }
-
-    const geoJsonFormat = new OlGeoJSONFormat();
-
-    let features: OlFeature[];
-    try {
-        features = geoJsonFormat.readFeatures(text);
-    } catch (e) {
-        try {
-            const geometry = geoJsonFormat.readGeometry(text);
-            features = [new OlFeature(geometry)];
-        } catch (e) {
-            throw new Error(i18n.get(`Invalid GeoJSON`));
-        }
-    }
-
-    const placeGroups: { [group: string]: PlaceGroup } = {};
-    let numPlaceGroups = 0;
-    let color = getUserPlaceColorName(0);
-    features.forEach(feature => {
-        const properties = feature.getProperties();
-        const geometry = feature.getGeometry();
-        if (geometry) {
-            let group = '';
-            let label = '';
-            const geoJsonProps: { [k: string]: number | string | boolean | null } = {color: 'red'};
-            if (properties) {
-                for (let propertyName in properties) {
-                    const propertyValue = properties[propertyName];
-                    if (propertyValue === null
-                        || typeof propertyValue === 'number'
-                        || typeof propertyValue === 'string'
-                        || typeof propertyValue === 'boolean') {
-                        geoJsonProps[propertyName] = propertyValue;
-                    }
-                    if (group === ''
-                        && typeof propertyValue === 'string'
-                        && groupNamesSet.has(propertyName.toLowerCase())) {
-                        group = propertyValue;
-                    }
-                    if (label === ''
-                        && typeof propertyValue === 'string'
-                        && labelNamesSet.has(propertyName.toLowerCase())) {
-                        label = propertyValue;
-                    }
-                }
-            }
-
-            if (group === '') {
-                const groupId = ++LAST_PLACE_GROUP_ID_GEO_JSON;
-                group = `${groupPrefix}-${groupId}`;
-            }
-            if (label === '') {
-                const labelId = ++LAST_PLACE_LABEL_ID_GEO_JSON;
-                label = `${labelPrefix}-${labelId}`;
-            }
-
-            let placeGroup = placeGroups[group];
-            if (!placeGroup) {
-                placeGroup = newUserPlaceGroup(group, []);
-                placeGroups[group] = placeGroup;
-                color = getUserPlaceColorName(numPlaceGroups);
-                numPlaceGroups++;
-            }
-
-            if (!geoJsonProps['color'])
-                geoJsonProps['color'] = color;
-            if (!geoJsonProps['label'])
-                geoJsonProps['label'] = label;
-            if (!geoJsonProps['source'])
-                geoJsonProps['source'] = 'GeoJSON';
-
-            placeGroup.features.push(newUserPlace(geometry, geoJsonProps));
-        }
-    });
-
-    return Object.getOwnPropertyNames(placeGroups).map(group => placeGroups[group]);
-}
-
-export function getUserPlacesFromWkt(text: string, options: WktOptions): PlaceGroup[] {
-
-    let groupPrefix = options.groupPrefix.trim();
-    if (groupPrefix === '') {
-        groupPrefix = defaultWktOptions.groupPrefix;
-    }
-    let group = options.group.trim();
-    if (group === '') {
-        const groupId = ++LAST_PLACE_GROUP_ID_WKT;
-        group = `${groupPrefix}${groupId}`;
-    }
-
-    let labelPrefix = options.labelPrefix.trim();
-    if (labelPrefix === '') {
-        labelPrefix = defaultWktOptions.labelPrefix;
-    }
-    let label = options.label.trim();
-    if (label === '') {
-        const labelId = ++LAST_PLACE_LABEL_ID_WKT;
-        label = `${labelPrefix}${labelId}`;
-    }
-
-    try {
-        const geometry = new OlWktFormat().readGeometry(text);
-        const geoJsonProps = {
-            color: getUserPlaceColorName(Math.floor(1000 * Math.random())),
-            label,
-            source: 'WKT'
-        };
-        const places = [newUserPlace(geometry, geoJsonProps)];
-        return [newUserPlaceGroup(group, places)];
-    } catch (e) {
-        throw new Error(i18n.get(`Invalid Geometry WKT`));
-    }
-}
-
-
-type ColumnNameToIndex = {[columnName: string]: number};
-
-function getColumnNameToIndex(columnNames: string[]): ColumnNameToIndex {
-    const indices: ColumnNameToIndex = {};
-    for (let index = 0; index < columnNames.length; index++) {
-        indices[columnNames[index].toLowerCase()] = index;
-    }
-    return indices;
-}
-
-
-function findColumnIndex(columnNameToIndex: ColumnNameToIndex, alternatives: string): number {
-    const alternativesLC = alternatives.toLowerCase().split(',').map(s => s.trim().toLowerCase());
-    for (let alternative of alternativesLC) {
-        const columnIndex: number | undefined = columnNameToIndex[alternative];
-        // noinspection SuspiciousTypeOfGuard
-        if (typeof columnIndex === 'number') {
-            return columnIndex;
-        }
-    }
-    return -1;
-}
-
