+++ src/selectors/dataSelectors.tsx
@@ -31,7 +31,6 @@ import { PlaceGroupTimeSeries, placeGroupToTimeSeries } from "../model/timeSerie
 export const datasetsSelector = (state: AppState) => state.dataState.datasets || [];
 export const colorBarsSelector = (state: AppState) => state.dataState.colorBars;
 export const timeSeriesGroupsSelector = (state: AppState) => state.dataState.timeSeriesGroups;
-export const userPlaceGroupSelector = (state: AppState) => state.dataState.userPlaceGroups[0];
 export const userPlaceGroupsSelector = (state: AppState) => state.dataState.userPlaceGroups;
 export const userServersSelector = (state: AppState) => state.dataState.userServers || [];
