+++ src/states/controlState.ts
@@ -31,9 +31,8 @@ import { Time, TimeRange } from '../model/timeSeries';
 import { loadUserSettings } from './userSettings';
 import { DEFAULT_MAP_CRS } from "../model/proj";
 import { CsvOptions, defaultCsvOptions } from "../model/user-place/csv";
-import { GeoJsonOptions, defaultGeoJsonOptions } from "../model/user-place/geojson";
-import { WktOptions, defaultWktOptions } from "../model/user-place/wkt";
-import { USER_DRAWING_PLACE_GROUP_ID } from "../model/place";
+import { defaultGeoJsonOptions, GeoJsonOptions } from "../model/user-place/geojson";
+import { defaultWktOptions, WktOptions } from "../model/user-place/wkt";
 
 
 export type TimeAnimationInterval = 250 | 500 | 1000 | 2500;
@@ -89,6 +88,7 @@ export interface ControlState {
     showTimeSeriesPointsOnly: boolean;
     showTimeSeriesErrorBars: boolean;
     showTimeSeriesMedian: boolean;
+    userDrawnPlaceGroupName: string,
     userPlacesFormatName: UserPlacesFormatName;
     userPlacesFormatOptions: UserPlacesFormatOptions;
     flyTo: OlGeometry | OlExtent | null;
@@ -122,7 +122,7 @@ export function newControlState(): ControlState {
     const state: ControlState = {
         selectedDatasetId: null,
         selectedVariableName: null,
-        selectedPlaceGroupIds: [USER_DRAWING_PLACE_GROUP_ID],
+        selectedPlaceGroupIds: [],
         selectedPlaceId: null,
         selectedUserPlaceId: null,
         selectedServerId: Config.instance.server.id,
@@ -136,6 +136,7 @@ export function newControlState(): ControlState {
         showTimeSeriesPointsOnly: false,
         showTimeSeriesErrorBars: true,
         showTimeSeriesMedian: branding.defaultAgg === 'median',
+        userDrawnPlaceGroupName: '',
         userPlacesFormatName: 'csv',
         userPlacesFormatOptions: {
             csv: {...defaultCsvOptions},
